requirements.txt
absl-py==1.3.0
ale-py==0.7.4
alembic==1.11.1
anyio==3.6.2
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
arrow==1.2.3
asttokens==2.2.1
attrs==22.2.0
autopep8==2.0.0
AutoROM==0.4.2
AutoROM.accept-rom-license==0.4.2
backcall==0.2.0
beautifulsoup4==4.11.1
bleach==5.0.1
cachetools==5.2.0
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
cloudpickle==2.2.0
cmaes==0.9.1
collision==1.2.2
coloredlogs==15.0.1
colorlog==6.7.0
comm==0.1.2
commonmark==0.9.1
contourpy==1.0.6
cycler==0.11.0
Cython==0.29.35
debugpy==1.6.5
decorator==5.1.1
defusedxml==0.7.1
dnspython==2.3.0
email-validator==2.0.0.post2
entrypoints==0.4
exceptiongroup==1.1.1
executing==1.2.0
fastapi==0.95.2
fastjsonschema==2.16.2
flatbuffers==23.5.26
fonttools==4.38.0
fqdn==1.5.1
gitdb==4.0.10
GitPython==3.1.31
glfw==2.5.5
google-auth==2.14.1
google-auth-oauthlib==0.4.6
greenlet==2.0.2
grpcio==1.50.0
gym==0.21.0
Gymnasium==0.26.3
gymnasium-notices==0.0.1
h11==0.14.0
httpcore==0.17.2
httptools==0.5.0
httpx==0.24.1
humanfriendly==10.0
idna==3.4
importlib-metadata==4.13.0
importlib-resources==5.10.0
iniconfig==2.0.0
ipykernel==6.20.2
ipython==8.8.0
ipython-genutils==0.2.0
ipywidgets==8.0.4
isoduration==20.11.0
itsdangerous==2.1.2
jedi==0.18.2
Jinja2==3.1.2
joblib==1.2.0
jsonpointer==2.3
jsonschema==4.17.3
jupyter==1.0.0
jupyter-console==6.4.4
jupyter-events==0.6.3
jupyter_client==7.4.9
jupyter_core==5.1.3
jupyter_server==2.1.0
jupyter_server_terminals==0.4.4
jupyterlab-pygments==0.2.2
jupyterlab-widgets==3.0.5
kiwisolver==1.4.4
lockfile==0.12.2
Mako==1.2.4
Markdown==3.4.1
MarkupSafe==2.1.1
matplotlib==3.6.2
matplotlib-inline==0.1.6
mistune==2.0.4
mpmath==1.2.1
nbclassic==0.4.8
nbclient==0.7.2
nbconvert==7.2.8
nbformat==5.7.3
nest-asyncio==1.5.6
notebook==6.5.2
notebook_shim==0.2.2
numpy==1.23.4
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
oauthlib==3.2.2
onnx==1.14.0
onnxruntime==1.15.0
opencv-python==4.6.0.66
optuna==3.1.1
orjson==3.9.0
packaging==21.3
pandas==1.5.1
pandocfilters==1.5.0
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.3.0
platformdirs==2.6.2
pluggy==1.0.0
prometheus-client==0.15.0
prompt-toolkit==3.0.36
protobuf==3.20.3
psutil==5.9.4
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.10.0
pycparser==2.21
pydantic==1.10.8
pygame==2.1.2
pyglet==2.0.0
Pygments==2.13.0
pyparsing==3.0.9
pyrsistent==0.19.3
pytest==7.3.1
python-dateutil==2.8.2
python-dotenv==1.0.0
python-json-logger==2.0.4
python-multipart==0.0.6
pytz==2022.6
PyYAML==6.0
pyzmq==25.0.0
qtconsole==5.4.0
QtPy==2.3.0
requests==2.28.1
requests-oauthlib==1.3.1
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==12.6.0
rsa==4.9
scikit-learn==1.2.0
scipy==1.9.3
seaborn==0.12.2
Send2Trash==1.8.0
six==1.16.0
smmap==5.0.0
sniffio==1.3.0
soupsieve==2.3.2.post1
SQLAlchemy==2.0.15
stable-baselines3==1.6.2
stack-data==0.6.2
starlette==0.27.0
swig==4.1.0
sympy==1.11.1
tensorboard==2.11.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
terminado==0.17.1
threadpoolctl==3.1.0
tinycss2==1.2.1
tomli==2.0.1
torch==1.13.0
tornado==6.2
tqdm==4.64.1
traitlets==5.8.1
typing_extensions==4.4.0
ujson==5.7.0
uri-template==1.2.0
urllib3==1.26.12
uvicorn==0.22.0
uvloop==0.17.0
watchfiles==0.19.0
wcwidth==0.2.6
webcolors==1.12
webencodings==0.5.1
websocket-client==1.4.2
websockets==11.0.3
Werkzeug==2.2.2
widgetsnbextension==4.0.5
zipp==3.10.0
